+++ mayan/apps/ocr/tests/test_views.py
@@ -7,7 +7,7 @@ from ..permissions import (
     permission_ocr_content_view, permission_ocr_document,
     permission_document_type_ocr_setup
 )
-from ..utils import get_document_ocr_content
+from ..utils import get_instance_ocr_content
 
 from .literals import TEST_DOCUMENT_CONTENT
 
@@ -184,7 +184,7 @@ class OCRViewsTestCase(OCRViewTestMixin, GenericDocumentViewTestCase):
 
         self.assert_download_response(
             response=response, content=(
-                ''.join(get_document_ocr_content(document=self.test_document))
+                ''.join(get_instance_ocr_content(document=self.test_document))
             ),
         )