+++ HISTORY.rst
@@ -12,6 +12,7 @@
 - Update the GitLab CI file to use the test makefile target and add
   migration testing.
 - Update the Docker run_tests command to perform migration testing.
+- Update translation files.
 
 3.3.9 (2020-01-18)
 ==================