+++ HISTORY.rst
@@ -24,6 +24,8 @@
 - Add command to show the current version of Mayan. The command is named
   ``showversion``. The command has one option `--build-string`` that will
   show the build string instead. Closes #MD-14.
+- Add command to check if the current version is the latest one. The command
+  is named ``checkversion``. Closes issue #MD-28.
 
 3.3.9 (2020-01-18)
 ==================